requirements.txt
Django==4.0.3
dj-database-url==0.5.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
djangorestframework==3.13.1
elasticsearch==7.17.2
psycopg2-binary==2.9.3

# Github requirements
git+https://github.com/iamjazzar/django-hitcount@65e01a6#egg=django-hitcount
